PCHR-2379: Adopt angularjs style guide for test files + fix precommit issues
