+++ org.civicrm.reqangular/karma.conf.js
@@ -1,13 +1,13 @@
 module.exports = function (config) {
-    var civicrmPath = '../../../../';
-    var civihrPath = 'tools/extensions/civihr/';
-    var extPath = civihrPath + 'org.civicrm.reqangular/';
+  var civicrmPath = '../../../../';
+  var civihrPath = 'tools/extensions/civihr/';
+  var extPath = civihrPath + 'org.civicrm.reqangular/';
 
-    config.set({
-        basePath: civicrmPath,
-        browsers: ['Chrome'],
-        frameworks: ['jasmine'],
-        files: [
+  config.set({
+    basePath: civicrmPath,
+    browsers: ['Chrome'],
+    frameworks: ['jasmine'],
+    files: [
             // the global dependencies
       'bower_components/jquery/dist/jquery.min.js',
       'bower_components/jquery-ui/jquery-ui.js',
@@ -15,12 +15,12 @@ module.exports = function (config) {
       'bower_components/select2/select2.min.js',
       'bower_components/jquery-validation/dist/jquery.validate.min.js',
       'packages/jquery/plugins/jquery.mousewheel.min.js',
-            'packages/jquery/plugins/jquery.blockUI.js',
-            'js/Common.js',
+      'packages/jquery/plugins/jquery.blockUI.js',
+      'js/Common.js',
 
             // manual loading of requirejs as to avoid interference with the global dependencies above
-            extPath + 'node_modules/requirejs/require.js',
-            extPath + 'node_modules/karma-requirejs/lib/adapter.js',
+      extPath + 'node_modules/requirejs/require.js',
+      extPath + 'node_modules/karma-requirejs/lib/adapter.js',
 
             // load test helpers
             { pattern: extPath + 'test/helpers/**/*.helper.js', included: true },
@@ -35,13 +35,13 @@ module.exports = function (config) {
             { pattern: extPath + 'test/mocks/**/*.js', included: false },
 
             // load tests
-            { pattern: extPath + 'test/**/*_test.js', included: false },
+            { pattern: extPath + 'test/**/*.spec.js', included: false },
 
             // the requireJS config file that bootstraps the whole test suite
-            extPath + 'test/test-main.js'
-        ],
-        exclude: [
-            extPath + 'src/common/angular-date/**/*.js'
-        ]
-    });
+      extPath + 'test/test-main.js'
+    ],
+    exclude: [
+      extPath + 'src/common/angular-date/**/*.js'
+    ]
+  });
 };
