+++ org.civicrm.reqangular/test/controllers/contact-actions/new-household-ctrl.spec.js
@@ -1,20 +1,20 @@
-/* eslint angular/di: 0 */
+/* eslint-env amd, jasmine */
 
 define([
   'common/angularMocks',
-  'common/controllers/contact-actions/new-organization-ctrl'
-], function() {
+  'common/controllers/contact-actions/new-household-ctrl'
+], function () {
   'use strict';
 
-  describe('NewOrganizationModalCtrl', function() {
+  describe('NewHouseholdModalCtrl', function () {
     var ctrl, $rootScope, $q, modalInstanceSpy, contactActionsStub, resultMock;
 
     beforeEach(module('common.apis', 'common.controllers'));
-    beforeEach(inject(function(_$controller_, _$rootScope_, _$q_) {
+    beforeEach(inject(function (_$controller_, _$rootScope_, _$q_) {
       $rootScope = _$rootScope_;
       $q = _$q_;
       initSpies();
-      ctrl = _$controller_('NewOrganizationModalCtrl', {
+      ctrl = _$controller_('NewHouseholdModalCtrl', {
         '$rootScope': $rootScope,
         '$uibModalInstance': modalInstanceSpy,
         'api.contactActions': contactActionsStub
@@ -25,17 +25,17 @@ define([
     /**
      * Jasmine spies initialization
      */
-    function initSpies() {
+    function initSpies () {
       modalInstanceSpy = jasmine.createSpyObj('modalInstanceSpy', ['dismiss']);
       contactActionsStub = {
-        save: jasmine.createSpyObj('saveSpy', ['newOrganization']),
-        getFormFields: jasmine.createSpyObj('getFormFieldsSpy', ['forNewOrganization'])
+        save: jasmine.createSpyObj('saveSpy', ['newHousehold']),
+        getFormFields: jasmine.createSpyObj('getFormFieldsSpy', ['forNewHousehold'])
       };
       resultMock = {
         test: true
       };
 
-      contactActionsStub.getFormFields.forNewOrganization.and.returnValue($q.resolve([{
+      contactActionsStub.getFormFields.forNewHousehold.and.returnValue($q.resolve([{
         id: '1',
         label: 'Field1',
         field_name: 'field1'
@@ -44,73 +44,73 @@ define([
         label: 'Field2',
         field_name: 'field2'
       }]));
-      contactActionsStub.save.newOrganization.and.returnValue($q.resolve(resultMock));
+      contactActionsStub.save.newHousehold.and.returnValue($q.resolve(resultMock));
     }
 
-    describe('cancel', function() {
-      it('closes the modal instance', function() {
+    describe('cancel', function () {
+      it('closes the modal instance', function () {
         ctrl.cancel();
         expect(modalInstanceSpy.dismiss).toHaveBeenCalled();
       });
     });
 
-    describe('submit', function() {
-      beforeEach(function() {
+    describe('submit', function () {
+      beforeEach(function () {
         ctrl.formFields[0].value = 'value1';
         ctrl.formFields[1].value = 'value2';
       });
 
-      describe('when there are no errors', function() {
-        beforeEach(function() {
+      describe('when there are no errors', function () {
+        beforeEach(function () {
           spyOn($rootScope, '$broadcast');
           ctrl.submit();
         });
 
-        it('saves the new organization', function() {
+        it('saves the new household', function () {
           $rootScope.$digest();
-          expect(contactActionsStub.save.newOrganization.calls.count()).toBe(1);
-          expect(contactActionsStub.save.newOrganization).toHaveBeenCalledWith({
+          expect(contactActionsStub.save.newHousehold.calls.count()).toBe(1);
+          expect(contactActionsStub.save.newHousehold).toHaveBeenCalledWith({
             field1: 'value1',
             field2: 'value2'
           });
         });
 
-        it('broadcasts the "newOrganizationCreated" event', function() {
+        it('broadcasts the "newHouseholdCreated" event', function () {
           $rootScope.$digest();
-          expect($rootScope.$broadcast).toHaveBeenCalledWith('newOrganizationCreated', resultMock);
+          expect($rootScope.$broadcast).toHaveBeenCalledWith('newHouseholdCreated', resultMock);
         });
 
-        it('doesn\'t set the error message', function() {
+        it('doesn\'t set the error message', function () {
           $rootScope.$digest();
           expect(ctrl.errorMsg.length).toBe(0);
         });
 
-        it('changes the "loading" property', function() {
+        it('changes the "loading" property', function () {
           expect(ctrl.loading).toBeTruthy();
           $rootScope.$digest();
           expect(ctrl.loading).toBeFalsy();
         });
       });
 
-      describe('when there are errors', function() {
-        beforeEach(function() {
-          contactActionsStub.save.newOrganization.and.returnValue($q.reject());
+      describe('when there are errors', function () {
+        beforeEach(function () {
+          contactActionsStub.save.newHousehold.and.returnValue($q.reject());
           spyOn($rootScope, '$broadcast');
           ctrl.submit();
         });
 
-        it('doesn\'t broadcast events', function() {
+        it('doesn\'t broadcast events', function () {
           $rootScope.$digest();
           expect($rootScope.$broadcast).not.toHaveBeenCalled();
         });
 
-        it('sets the error message', function() {
+        it('sets the error message', function () {
           $rootScope.$digest();
-          expect(contactActionsStub.save.newOrganization.calls.count()).toBe(1);
+          expect(contactActionsStub.save.newHousehold.calls.count()).toBe(1);
           expect(ctrl.errorMsg.length).not.toBe(0);
         });
 
-        it('changes the "loading" property', function() {
+        it('changes the "loading" property', function () {
           expect(ctrl.loading).toBeTruthy();
           $rootScope.$digest();
           expect(ctrl.loading).toBeFalsy();
