+++ org.civicrm.reqangular/test/controllers/contact-actions/new-organization-ctrl.spec.js
@@ -1,12 +1,12 @@
-/* eslint angular/di: 0 */
+/* eslint-env amd, jasmine */
 
 define([
   'common/angularMocks',
-  'common/controllers/contact-actions/new-household-ctrl'
+  'common/controllers/contact-actions/new-organization-ctrl'
 ], function () {
   'use strict';
 
-  describe('NewHouseholdModalCtrl', function () {
+  describe('NewOrganizationModalCtrl', function () {
     var ctrl, $rootScope, $q, modalInstanceSpy, contactActionsStub, resultMock;
 
     beforeEach(module('common.apis', 'common.controllers'));
@@ -14,7 +14,7 @@ define([
       $rootScope = _$rootScope_;
       $q = _$q_;
       initSpies();
-      ctrl = _$controller_('NewHouseholdModalCtrl', {
+      ctrl = _$controller_('NewOrganizationModalCtrl', {
         '$rootScope': $rootScope,
         '$uibModalInstance': modalInstanceSpy,
         'api.contactActions': contactActionsStub
@@ -25,27 +25,27 @@ define([
     /**
      * Jasmine spies initialization
      */
-     function initSpies() {
-       modalInstanceSpy = jasmine.createSpyObj('modalInstanceSpy', ['dismiss']);
-       contactActionsStub = {
-         save: jasmine.createSpyObj('saveSpy', ['newHousehold']),
-         getFormFields: jasmine.createSpyObj('getFormFieldsSpy', ['forNewHousehold'])
-       };
-       resultMock = {
-         test: true
-       };
+    function initSpies () {
+      modalInstanceSpy = jasmine.createSpyObj('modalInstanceSpy', ['dismiss']);
+      contactActionsStub = {
+        save: jasmine.createSpyObj('saveSpy', ['newOrganization']),
+        getFormFields: jasmine.createSpyObj('getFormFieldsSpy', ['forNewOrganization'])
+      };
+      resultMock = {
+        test: true
+      };
 
-       contactActionsStub.getFormFields.forNewHousehold.and.returnValue($q.resolve([{
-         id: '1',
-         label: 'Field1',
-         field_name: 'field1'
-       }, {
-         id: '2',
-         label: 'Field2',
-         field_name: 'field2'
-       }]));
-       contactActionsStub.save.newHousehold.and.returnValue($q.resolve(resultMock));
-     }
+      contactActionsStub.getFormFields.forNewOrganization.and.returnValue($q.resolve([{
+        id: '1',
+        label: 'Field1',
+        field_name: 'field1'
+      }, {
+        id: '2',
+        label: 'Field2',
+        field_name: 'field2'
+      }]));
+      contactActionsStub.save.newOrganization.and.returnValue($q.resolve(resultMock));
+    }
 
     describe('cancel', function () {
       it('closes the modal instance', function () {
@@ -66,18 +66,18 @@ define([
           ctrl.submit();
         });
 
-        it('saves the new household', function () {
+        it('saves the new organization', function () {
           $rootScope.$digest();
-          expect(contactActionsStub.save.newHousehold.calls.count()).toBe(1);
-          expect(contactActionsStub.save.newHousehold).toHaveBeenCalledWith({
+          expect(contactActionsStub.save.newOrganization.calls.count()).toBe(1);
+          expect(contactActionsStub.save.newOrganization).toHaveBeenCalledWith({
             field1: 'value1',
             field2: 'value2'
           });
         });
 
-        it('broadcasts the "newHouseholdCreated" event', function () {
+        it('broadcasts the "newOrganizationCreated" event', function () {
           $rootScope.$digest();
-          expect($rootScope.$broadcast).toHaveBeenCalledWith('newHouseholdCreated', resultMock);
+          expect($rootScope.$broadcast).toHaveBeenCalledWith('newOrganizationCreated', resultMock);
         });
 
         it('doesn\'t set the error message', function () {
@@ -94,7 +94,7 @@ define([
 
       describe('when there are errors', function () {
         beforeEach(function () {
-          contactActionsStub.save.newHousehold.and.returnValue($q.reject());
+          contactActionsStub.save.newOrganization.and.returnValue($q.reject());
           spyOn($rootScope, '$broadcast');
           ctrl.submit();
         });
@@ -106,7 +106,7 @@ define([
 
         it('sets the error message', function () {
           $rootScope.$digest();
-          expect(contactActionsStub.save.newHousehold.calls.count()).toBe(1);
+          expect(contactActionsStub.save.newOrganization.calls.count()).toBe(1);
           expect(ctrl.errorMsg.length).not.toBe(0);
         });
