+++ org.civicrm.reqangular/test/mocks/models/instances/contact-instance-mock.js
@@ -1,36 +1,37 @@
+/* eslint-env amd */
+
 define([
-    'common/lodash',
-    'common/mocks/module'
+  'common/lodash',
+  'common/mocks/module'
 ], function (_, mocks) {
-    'use strict';
+  'use strict';
 
-    mocks.factory('ContactInstanceMock', ['$q', 'ContactInstance', function ($q, instance) {
+  mocks.factory('ContactInstanceMock', ['$q', 'ContactInstance', function ($q, instance) {
+    return _.assign(Object.create(instance), {
 
-        return _.assign(Object.create(instance), {
+      /**
+       * Checks if the given object is a modal instance
+       *
+       * @param {object} object
+       * @return {boolean}
+       */
+      isInstance: function (object) {
+        return _.isEqual(_.functions(object), _.functions(instance));
+      },
 
-            /**
-             * Checks if the given object is a modal instance
-             *
-             * @param {object} object
-             * @return {boolean}
-             */
-            isInstance: function (object) {
-                return _.isEqual(_.functions(object), _.functions(instance));
-            }
-        });
+      /**
+      * Returns a promise that will resolve with the given value
+      *
+      * @param {any} value
+      * @return {Promise}
+      */
+      promiseResolvedWit: function (value) {
+        var deferred = $q.defer();
 
+        deferred.resolve(value);
 
-        /**
-         * Returns a promise that will resolve with the given value
-         *
-         * @param {any} value
-         * @return {Promise}
-         */
-        function promiseResolvedWith(value) {
-            var deferred = $q.defer();
-            deferred.resolve(value);
-
-            return deferred.promise;
-        }
-    }]);
+        return deferred.promise;
+      }
+    });
+  }]);
 });
