+++ org.civicrm.reqangular/test/mocks/module.js
@@ -1,7 +1,9 @@
-define([
-    'common/angular'
-], function (angular) {
-    'use strict';
+/* eslint-env amd */
 
-    return angular.module('common.mocks', []);
+define([
+  'common/angular'
+], function (angular) {
+  'use strict';
+
+  return angular.module('common.mocks', []);
 });
