+++ org.civicrm.reqangular/test/mocks/services/hr-settings-mock.js
@@ -1,11 +1,13 @@
-define([
-    'common/mocks/module'
-], function (mocks) {
-    'use strict';
+/* eslint-env amd */
 
-    mocks.factory('HR_settingsMock', function () {
-        return {
-            DATE_FORMAT: 'dd/MM/yyyy'
-        };
-    });
+define([
+  'common/mocks/module'
+], function (mocks) {
+  'use strict';
+
+  mocks.factory('HR_settingsMock', function () {
+    return {
+      DATE_FORMAT: 'dd/MM/yyyy'
+    };
+  });
 });
