+++ org.civicrm.reqangular/test/services/contact-actions/contact-actions.spec.js
@@ -1,20 +1,21 @@
 /* eslint angular/di: 0, jasmine/no-spec-dupes: 0 */
+/* eslint-env amd, jasmine */
 
 define([
   'common/angularMocks',
   'common/services/api/contact-actions'
-], function() {
+], function () {
   'use strict';
 
-  describe('ContactActions API', function() {
+  describe('ContactActions API', function () {
     var $rootScope, apiSpy;
 
-    beforeEach(module('common.apis', function($provide) {
+    beforeEach(module('common.apis', function ($provide) {
       apiSpy = jasmine.createSpyObj('apiSpy', ['extend', 'sendPOST', 'sendGET']);
       $provide.value('api', apiSpy);
     }));
 
-    beforeEach(inject(function($injector, $q, _$rootScope_) {
+    beforeEach(inject(function ($injector, $q, _$rootScope_) {
       $rootScope = _$rootScope_;
       apiSpy.extend.and.returnValue({});
       apiSpy.sendPOST.and.returnValue($q.resolve({
@@ -28,7 +29,7 @@ define([
       $rootScope.$apply();
     }));
 
-    it('calls api.extend with correct parameters', function() {
+    it('calls api.extend with correct parameters', function () {
       expect(apiSpy.extend.calls.count()).toBe(1);
       expect(apiSpy.extend.calls.mostRecent().args.length).toBe(1);
       expect('getOptions' in apiSpy.extend.calls.mostRecent().args[0]).toBeTruthy();
@@ -51,21 +52,21 @@ define([
       expect('forNewHousehold' in apiSpy.extend.calls.mostRecent().args[0].getFormFields).toBeTruthy();
     });
 
-    describe('getOptions.forContactType', function() {
+    describe('getOptions.forContactType', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forContactType.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['Contact', 'getoptions', {
           field: 'contact_type',
@@ -74,21 +75,21 @@ define([
       });
     });
 
-    describe('getOptions.forGroup', function() {
+    describe('getOptions.forGroup', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forGroup.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['GroupContact', 'getoptions', {
           field: 'group_id',
@@ -97,21 +98,21 @@ define([
       });
     });
 
-    describe('getOptions.forTag', function() {
+    describe('getOptions.forTag', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forTag.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['EntityTag', 'getoptions', {
           field: 'tag_id',
@@ -120,21 +121,21 @@ define([
       });
     });
 
-    describe('getOptions.forStateProvince', function() {
+    describe('getOptions.forStateProvince', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forStateProvince.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['Address', 'getoptions', {
           field: 'state_province_id',
@@ -143,21 +144,21 @@ define([
       });
     });
 
-    describe('getOptions.forCountry', function() {
+    describe('getOptions.forCountry', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forCountry.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['Address', 'getoptions', {
           field: 'country_id',
@@ -166,21 +167,21 @@ define([
       });
     });
 
-    describe('getOptions.forGender', function() {
+    describe('getOptions.forGender', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forGender.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['Contact', 'getoptions', {
           field: 'gender_id',
@@ -189,21 +190,21 @@ define([
       });
     });
 
-    describe('getOptions.forDeceased', function() {
+    describe('getOptions.forDeceased', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getOptions.forDeceased.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['Contact', 'getoptions', {
           field: 'is_deceased',
@@ -212,25 +213,25 @@ define([
       });
     });
 
-    describe('save.newIndividual', function() {
-      describe('when the newIndividual data doesn\'t have the "email" field', function() {
+    describe('save.newIndividual', function () {
+      describe('when the newIndividual data doesn\'t have the "email" field', function () {
         var result;
-        beforeEach(function() {
+        beforeEach(function () {
           result = apiSpy.extend.calls.mostRecent().args[0].save.newIndividual.call(apiSpy, {
             first_name: 'First Name',
             last_name: 'Last Name'
           });
         });
 
-        it('returns the correct result', function(done) {
-          result.then(function(data) {
+        it('returns the correct result', function (done) {
+          result.then(function (data) {
             expect(data).toEqual('values from post');
             done();
           });
           $rootScope.$apply();
         });
 
-        it('calls api.sendPOST correctly', function() {
+        it('calls api.sendPOST correctly', function () {
           $rootScope.$apply();
           expect(apiSpy.sendPOST.calls.count()).toBe(1);
           expect(apiSpy.sendPOST.calls.argsFor(0)).toEqual(['Contact', 'create', {
@@ -240,15 +241,15 @@ define([
           }]);
         });
 
-        it('doesn\'t call api.sendGET', function() {
+        it('doesn\'t call api.sendGET', function () {
           $rootScope.$apply();
           expect(apiSpy.sendGET.calls.count()).toBe(0);
         });
       });
 
-      describe('when the newIndividual data has the "email" field', function() {
+      describe('when the newIndividual data has the "email" field', function () {
         var result;
-        beforeEach(function() {
+        beforeEach(function () {
           result = apiSpy.extend.calls.mostRecent().args[0].save.newIndividual.call(apiSpy, {
             first_name: 'First Name',
             last_name: 'Last Name',
@@ -256,15 +257,15 @@ define([
           });
         });
 
-        it('returns the correct result', function(done) {
-          result.then(function(data) {
+        it('returns the correct result', function (done) {
+          result.then(function (data) {
             expect(data).toEqual('values from post');
             done();
           });
           $rootScope.$apply();
         });
 
-        it('calls api.sendPOST correctly', function() {
+        it('calls api.sendPOST correctly', function () {
           $rootScope.$apply();
           expect(apiSpy.sendPOST.calls.count()).toBe(1);
           expect(apiSpy.sendPOST.calls.argsFor(0)).toEqual(['Contact', 'create', {
@@ -275,7 +276,7 @@ define([
           }]);
         });
 
-        it('calls api.sendGET', function() {
+        it('calls api.sendGET', function () {
           $rootScope.$apply();
           expect(apiSpy.sendGET.calls.count()).toBe(1);
           expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['CustomField', 'get', {
@@ -287,24 +288,24 @@ define([
       });
     });
 
-    describe('save.newOrganization', function() {
-      describe('when the newOrganization data doesn\'t have the "email" field', function() {
+    describe('save.newOrganization', function () {
+      describe('when the newOrganization data doesn\'t have the "email" field', function () {
         var result;
-        beforeEach(function() {
+        beforeEach(function () {
           result = apiSpy.extend.calls.mostRecent().args[0].save.newOrganization.call(apiSpy, {
             organization_name: 'Organization Name'
           });
         });
 
-        it('returns the correct result', function(done) {
-          result.then(function(data) {
+        it('returns the correct result', function (done) {
+          result.then(function (data) {
             expect(data).toEqual('values from post');
             done();
           });
           $rootScope.$apply();
         });
 
-        it('calls api.sendPOST correctly', function() {
+        it('calls api.sendPOST correctly', function () {
           $rootScope.$apply();
           expect(apiSpy.sendPOST.calls.count()).toBe(1);
           expect(apiSpy.sendPOST.calls.argsFor(0)).toEqual(['Contact', 'create', {
@@ -313,30 +314,30 @@ define([
           }]);
         });
 
-        it('doesn\'t call api.sendGET', function() {
+        it('doesn\'t call api.sendGET', function () {
           $rootScope.$apply();
           expect(apiSpy.sendGET.calls.count()).toBe(0);
         });
       });
 
-      describe('when the newOrganization data has the "email" field', function() {
+      describe('when the newOrganization data has the "email" field', function () {
         var result;
-        beforeEach(function() {
+        beforeEach(function () {
           result = apiSpy.extend.calls.mostRecent().args[0].save.newOrganization.call(apiSpy, {
             organization_name: 'Organization Name',
             email: 'x@x.com'
           });
         });
 
-        it('returns the correct result', function(done) {
-          result.then(function(data) {
+        it('returns the correct result', function (done) {
+          result.then(function (data) {
             expect(data).toEqual('values from post');
             done();
           });
           $rootScope.$apply();
         });
 
-        it('calls api.sendPOST correctly', function() {
+        it('calls api.sendPOST correctly', function () {
           $rootScope.$apply();
           expect(apiSpy.sendPOST.calls.count()).toBe(1);
           expect(apiSpy.sendPOST.calls.argsFor(0)).toEqual(['Contact', 'create', {
@@ -346,7 +347,7 @@ define([
           }]);
         });
 
-        it('calls api.sendGET', function() {
+        it('calls api.sendGET', function () {
           $rootScope.$apply();
           expect(apiSpy.sendGET.calls.count()).toBe(1);
           expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['CustomField', 'get', {
@@ -358,24 +359,24 @@ define([
       });
     });
 
-    describe('save.newHousehold', function() {
-      describe('when the newHousehold data doesn\'t have the "email" field', function() {
+    describe('save.newHousehold', function () {
+      describe('when the newHousehold data doesn\'t have the "email" field', function () {
         var result;
-        beforeEach(function() {
+        beforeEach(function () {
           result = apiSpy.extend.calls.mostRecent().args[0].save.newHousehold.call(apiSpy, {
             household_name: 'Household Name'
           });
         });
 
-        it('returns the correct result', function(done) {
-          result.then(function(data) {
+        it('returns the correct result', function (done) {
+          result.then(function (data) {
             expect(data).toEqual('values from post');
             done();
           });
           $rootScope.$apply();
         });
 
-        it('calls api.sendPOST correctly', function() {
+        it('calls api.sendPOST correctly', function () {
           $rootScope.$apply();
           expect(apiSpy.sendPOST.calls.count()).toBe(1);
           expect(apiSpy.sendPOST.calls.argsFor(0)).toEqual(['Contact', 'create', {
@@ -384,30 +385,30 @@ define([
           }]);
         });
 
-        it('doesn\'t call api.sendGET', function() {
+        it('doesn\'t call api.sendGET', function () {
           $rootScope.$apply();
           expect(apiSpy.sendGET.calls.count()).toBe(0);
         });
       });
 
-      describe('when the newHousehold data has the "email" field', function() {
+      describe('when the newHousehold data has the "email" field', function () {
         var result;
-        beforeEach(function() {
+        beforeEach(function () {
           result = apiSpy.extend.calls.mostRecent().args[0].save.newHousehold.call(apiSpy, {
             household_name: 'Household Name',
             email: 'x@x.com'
           });
         });
 
-        it('returns the correct result', function(done) {
-          result.then(function(data) {
+        it('returns the correct result', function (done) {
+          result.then(function (data) {
             expect(data).toEqual('values from post');
             done();
           });
           $rootScope.$apply();
         });
 
-        it('calls api.sendPOST correctly', function() {
+        it('calls api.sendPOST correctly', function () {
           $rootScope.$apply();
           expect(apiSpy.sendPOST.calls.count()).toBe(1);
           expect(apiSpy.sendPOST.calls.argsFor(0)).toEqual(['Contact', 'create', {
@@ -417,7 +418,7 @@ define([
           }]);
         });
 
-        it('calls api.sendGET', function() {
+        it('calls api.sendGET', function () {
           $rootScope.$apply();
           expect(apiSpy.sendGET.calls.count()).toBe(1);
           expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['CustomField', 'get', {
@@ -429,21 +430,21 @@ define([
       });
     });
 
-    describe('getFormFields.forNewIndividual', function() {
+    describe('getFormFields.forNewIndividual', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getFormFields.forNewIndividual.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['UFField', 'get', {
           uf_group_id: 'new_individual',
@@ -452,21 +453,21 @@ define([
       });
     });
 
-    describe('getFormFields.forNewOrganization', function() {
+    describe('getFormFields.forNewOrganization', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getFormFields.forNewOrganization.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['UFField', 'get', {
           uf_group_id: 'new_organization',
@@ -475,21 +476,21 @@ define([
       });
     });
 
-    describe('getFormFields.forNewHousehold', function() {
+    describe('getFormFields.forNewHousehold', function () {
       var result;
-      beforeEach(function() {
+      beforeEach(function () {
         result = apiSpy.extend.calls.mostRecent().args[0].getFormFields.forNewHousehold.call(apiSpy);
       });
 
-      it('returns the correct result', function(done) {
-        result.then(function(data) {
+      it('returns the correct result', function (done) {
+        result.then(function (data) {
           expect(data).toEqual(['values from get']);
           done();
         });
         $rootScope.$apply();
       });
 
-      it('calls api.sendGET correctly', function() {
+      it('calls api.sendGET correctly', function () {
         expect(apiSpy.sendGET.calls.count()).toBe(1);
         expect(apiSpy.sendGET.calls.argsFor(0)).toEqual(['UFField', 'get', {
           uf_group_id: 'new_household',
