+++ org.civicrm.reqangular/test/services/pub-sub.spec.js
@@ -1,3 +1,6 @@
+/* globals angular, inject */
+/* eslint-env amd, jasmine */
+
 define([
   'common/angular',
   'common/angularMocks',
@@ -7,26 +10,26 @@ define([
 
   describe('Pub Sub Service test', function () {
     var ctrlConstructor, $rootScope, $timeout;
-    beforeEach(function(){
-      var moduleA = angular.module("moduleA", ["common.services"]);
-      moduleA.controller("controllerA",
-        function($scope, pubSub) {
-          $scope.publish = function(data) {
-            pubSub.publish("changeB", data);
+    beforeEach(function () {
+      var moduleA = angular.module('moduleA', ['common.services']);
+      moduleA.controller('controllerA',
+        function ($scope, pubSub) {
+          $scope.publish = function (data) {
+            pubSub.publish('changeB', data);
           };
         }
       );
 
-      var moduleB = angular.module("moduleB", ["common.services"]);
-      moduleB.controller("controllerB",
-        function($scope, pubSub) {
-          $scope.subscriber = pubSub.subscribe("changeB", function(data) {
+      var moduleB = angular.module('moduleB', ['common.services']);
+      moduleB.controller('controllerB',
+        function ($scope, pubSub) {
+          $scope.subscriber = pubSub.subscribe('changeB', function (data) {
             $scope.data = data;
           });
         }
       );
 
-      module('moduleA', 'moduleB')
+      module('moduleA', 'moduleB');
     });
 
     beforeEach(inject(function (_$controller_, _$rootScope_, _$timeout_) {
@@ -40,10 +43,10 @@ define([
       var $scopeB = $rootScope.$new();
       ctrlConstructor('controllerA', { $scope: $scopeA });
       ctrlConstructor('controllerB', { $scope: $scopeB });
-      $scopeA.publish("someval");
-      $timeout(function(){
-        expect($scopeB.data).toBe("someval");
-      },0);
+      $scopeA.publish('someval');
+      $timeout(function () {
+        expect($scopeB.data).toBe('someval');
+      }, 0);
       $timeout.flush();
     });
 
@@ -53,10 +56,10 @@ define([
       ctrlConstructor('controllerA', { $scope: $scopeA });
       ctrlConstructor('controllerB', { $scope: $scopeB });
       $scopeB.subscriber.remove();
-      $scopeA.publish("someval");
-      $timeout(function(){
-        expect($scopeB.data).not.toBe("someval");
-      },0);
+      $scopeA.publish('someval');
+      $timeout(function () {
+        expect($scopeB.data).not.toBe('someval');
+      }, 0);
       $timeout.flush();
     });
   });
